feat: merge footer to header, design styling
