--- sdk-app/src/Footer.module.scss
@@ -1,138 +1,0 @@
-.root {
-  display: flex;
-  align-items: center;
-  height: 3rem;
-  background: #ffffff;
-  color: var(--color-text);
-  padding: 0 1rem;
-  gap: 1rem;
-  flex-shrink: 0;
-  font-size: 0.8125rem;
-  z-index: 10;
-  border-top: 1px solid var(--color-border);
-  box-shadow: 0px -2px 5px 0px rgba(0, 0, 0, 0.05);
-}
-
-.divider {
-  width: 0.0625rem;
-  height: 1.5rem;
-  background: var(--color-border);
-}
-
-.badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.375rem;
-  padding: 0.1875rem 0.625rem;
-  border-radius: 0.25rem;
-  font-size: 0.6875rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.03125rem;
-}
-
-.badgeLabel {
-  font-weight: 400;
-  opacity: 0.6;
-  font-size: 0.625rem;
-}
-
-.badgeEnv {
-  background: rgba(13, 110, 253, 0.1);
-  color: #0d6efd;
-}
-
-.badgeBuild {
-  background: rgba(25, 135, 84, 0.1);
-  color: #198754;
-}
-
-.badgeDemoType {
-  background: rgba(255, 193, 7, 0.15);
-  color: #997404;
-}
-
-.company {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-family: monospace;
-  font-size: 0.75rem;
-  color: var(--color-text-muted);
-  max-width: 16.25rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-
-  .badgeLabel {
-    font-family:
-      -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
-    font-size: 0.625rem;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 0.03125rem;
-  }
-}
-
-.tokenStatus {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-size: 0.75rem;
-  color: var(--color-text-muted);
-}
-
-.tokenDot {
-  width: 0.5rem;
-  height: 0.5rem;
-  border-radius: 50%;
-}
-
-.tokenDotValid {
-  background: var(--color-success);
-}
-
-.tokenDotExpired {
-  background: var(--color-danger);
-}
-
-.tokenDotChecking {
-  background: var(--color-warning);
-  animation: pulse 1s infinite;
-}
-
-.tokenDotUnknown {
-  background: var(--color-text-muted);
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.4;
-  }
-}
-
-.spacer {
-  flex: 1;
-}
-
-.gear {
-  background: none;
-  border: 1px solid var(--color-border);
-  color: var(--color-text);
-  cursor: pointer;
-  padding: 0.375rem 0.625rem;
-  border-radius: 0.375rem;
-  font-size: 0.8125rem;
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  transition: background 0.15s;
-
-  &:hover {
-    background: var(--color-hover-bg);
-  }
-}
