+++ sdk-app/src/ModeSwitcher.module.scss
@@ -5,6 +5,9 @@
 }
 
 .option {
+  display: inline-flex;
+  gap: 0.375rem;
+  align-items: center;
   padding: 0.6rem 0.75rem;
   font-size: 0.875rem;
   font-weight: 600;
@@ -12,6 +15,7 @@
   color: var(--color-text);
   background: var(--color-bg);
   border: 1px solid var(--color-border);
+  max-height: 2.3rem;
   transition:
     background 0.15s,
     color 0.15s;
@@ -28,6 +32,11 @@
   &:hover {
     background: var(--color-hover-bg);
     color: var(--color-text);
+  }
+
+  svg {
+    width: 1.2rem;
+    height: 1.2rem;
   }
 }
