+++ sdk-app/src/Sidebar.module.scss
@@ -44,13 +44,10 @@
   justify-content: space-between;
   padding: 0.5rem 1rem;
   cursor: pointer;
-  font-weight: 600;
-  font-size: 0.75rem;
-  text-transform: uppercase;
-  letter-spacing: 0.03125rem;
+  font-weight: 500;
+  font-size: 0.8rem;
   color: var(--color-text-muted);
   user-select: none;
-
   &:hover {
     color: var(--color-text);
   }
@@ -64,14 +61,18 @@
   font-weight: 500;
 }
 
-.categoryArrow {
+.categoryTitle {
   display: inline-flex;
-  font-size: 0.625rem;
+  align-items: center;
+  gap: 0.375rem;
+  font-weight: 500;
+  font-size: 0.8rem;
+  color: var(--color-text-muted);
+}
+
+.categoryArrow {
   transition: transform 0.15s;
-  margin-right: 0.375rem;
   transform: rotate(90deg);
-  position: relative;
-  top: toRem(5);
 }
 
 .categoryArrowCollapsed {
