+++ sdk-app/src/TopBar.module.scss
@@ -1,6 +1,7 @@
 .root {
   display: flex;
   align-items: center;
+  justify-content: space-between;
   background: #ffffff;
   color: var(--color-text);
   padding: 0.75rem 1rem;
@@ -16,6 +17,7 @@
   font-weight: 600;
   font-size: 1.1rem;
   color: var(--color-text);
+  white-space: nowrap;
 }
 
 .designTitle {
@@ -25,6 +27,149 @@
   margin-left: 0.25rem;
 }
 
-.spacer {
-  flex: 1;
+.rightBar {
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+}
+
+.envInfo {
+  display: flex;
+  align-items: center;
+  gap: 0.625rem;
+}
+
+.actions {
+  display: flex;
+  align-items: center;
+  gap: 0.625rem;
+}
+
+.badge {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.375rem;
+  padding: 0.1875rem 0.625rem;
+  border-radius: 0.25rem;
+  font-size: 0.6875rem;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.03125rem;
+}
+
+.badgeLabel {
+  font-weight: 400;
+  opacity: 0.6;
+  font-size: 0.625rem;
+}
+
+.badgeEnv {
+  background: rgba(13, 110, 253, 0.1);
+  color: #0d6efd;
+}
+
+.badgeBuild {
+  background: rgba(25, 135, 84, 0.1);
+  color: #198754;
+}
+
+.badgeDemoType {
+  background: rgba(255, 193, 7, 0.15);
+  color: #997404;
+}
+
+.divider {
+  width: 0.0625rem;
+  height: 1.5rem;
+  background: var(--color-border);
+}
+
+.company {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.375rem;
+  font-family: monospace;
+  font-size: 0.75rem;
+  color: var(--color-text-muted);
+  max-width: 16.25rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+
+  .badgeLabel {
+    font-family:
+      -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
+    font-size: 0.625rem;
+    font-weight: 600;
+    text-transform: uppercase;
+    letter-spacing: 0.03125rem;
+  }
+}
+
+.tokenStatus {
+  display: flex;
+  align-items: center;
+  gap: 0.375rem;
+  font-size: 0.75rem;
+  color: var(--color-text-muted);
+}
+
+.tokenDot {
+  width: 0.5rem;
+  height: 0.5rem;
+  border-radius: 50%;
+}
+
+.tokenDotValid {
+  background: var(--color-success);
+}
+
+.tokenDotExpired {
+  background: var(--color-danger);
+}
+
+.tokenDotChecking {
+  background: var(--color-warning);
+  animation: pulse 1s infinite;
+}
+
+.tokenDotUnknown {
+  background: var(--color-text-muted);
+}
+
+@keyframes pulse {
+  0%,
+  100% {
+    opacity: 1;
+  }
+  50% {
+    opacity: 0.4;
+  }
+}
+
+.settingsBtn {
+  font-family: inherit;
+  background: var(--color-bg);
+  border: 1px solid var(--color-border);
+  color: var(--color-text);
+  cursor: pointer;
+  padding: 0.6rem 0.75rem;
+  max-height: 2.3rem;
+  border-radius: 0.375rem;
+  font-size: 0.875rem;
+  font-weight: 600;
+  display: flex;
+  align-items: center;
+  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.08);
+  gap: 0.375rem;
+  transition: background 0.15s;
+
+  svg {
+    width: 1.2rem;
+    height: 1.2rem;
+  }
+
+  &:hover {
+    background: var(--color-hover-bg);
+  }
 }
